+++ Sources/SpanWallpaper/Rotation.swift
@@ -166,6 +166,7 @@ class RotationManager {
         } catch {
             Log.info("Rotation error: \(error)")
         }
+        purgeCache()
     }
 
     func applyPrevious() {
@@ -187,6 +188,7 @@ class RotationManager {
         } catch {
             Log.info("Previous error: \(error)")
         }
+        purgeCache()
     }
 
     func reapplyCurrent() {
@@ -261,6 +263,15 @@ class RotationManager {
                 Log.info("Reapplied single image: \(singlePath)")
             } catch {
                 Log.info("Reapply error: \(error)")
+            }
+        }
+    }
+
+    private func purgeCache() {
+        if WallpaperCache.hasAccess() {
+            let deleted = WallpaperCache.purge(keeping: 10)
+            if deleted > 0 {
+                Log.info("Purged \(deleted) macOS wallpaper cache files")
             }
         }
     }